pyproject.toml
[project]
name = "breakfix"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Roberto Abdelkader Martínez Pérez", email = "robertomartinezp@gmail.com" }]
readme = "README.md"
urls = { Repository = "https://github.com/nilp0inter/breakfix" }
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.25",
    "cosmic-ray>=8.0",
    "prefect>=3.0",
    "pydantic-ai[prefect]>=1.47.0",
    "pydantic-ai-filesystem-sandbox>=0.9.0",
    "pydeps>=3.0.2",
    "pyscaffold>=4.6",
]

[project.scripts]
breakfix = "breakfix.cli:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
